+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/HttpCompilerPluginUtil.java
@@ -20,6 +20,7 @@ package io.ballerina.stdlib.http.compiler;
 
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.api.Types;
+import io.ballerina.compiler.api.symbols.ClassSymbol;
 import io.ballerina.compiler.api.symbols.FunctionSymbol;
 import io.ballerina.compiler.api.symbols.FunctionTypeSymbol;
 import io.ballerina.compiler.api.symbols.IntersectionTypeSymbol;
@@ -32,13 +33,16 @@ import io.ballerina.compiler.api.symbols.TypeReferenceTypeSymbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
 import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
 import io.ballerina.compiler.syntax.tree.AnnotationNode;
+import io.ballerina.compiler.syntax.tree.ClassDefinitionNode;
 import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
+import io.ballerina.compiler.syntax.tree.MethodDeclarationNode;
 import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.compiler.syntax.tree.NodeList;
 import io.ballerina.compiler.syntax.tree.ObjectTypeDescriptorNode;
 import io.ballerina.compiler.syntax.tree.ReturnTypeDescriptorNode;
 import io.ballerina.compiler.syntax.tree.ServiceDeclarationNode;
 import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.projects.Document;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticFactory;
@@ -47,12 +51,19 @@ import io.ballerina.tools.diagnostics.DiagnosticProperty;
 import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 import io.ballerina.tools.diagnostics.Location;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 
+import static io.ballerina.compiler.api.symbols.SymbolKind.TYPE_DEFINITION;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.ANNOTATION;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.AT_TOKEN;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.CLASS_DEFINITION;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.RESOURCE_ACCESSOR_DECLARATION;
+import static io.ballerina.compiler.syntax.tree.SyntaxKind.RESOURCE_ACCESSOR_DEFINITION;
 import static io.ballerina.stdlib.http.compiler.Constants.ANYDATA;
 import static io.ballerina.stdlib.http.compiler.Constants.ARRAY_OF_MAP_OF_ANYDATA;
 import static io.ballerina.stdlib.http.compiler.Constants.BALLERINA;
@@ -60,8 +71,10 @@ import static io.ballerina.stdlib.http.compiler.Constants.BOOLEAN;
 import static io.ballerina.stdlib.http.compiler.Constants.BOOLEAN_ARRAY;
 import static io.ballerina.stdlib.http.compiler.Constants.BYTE_ARRAY;
 import static io.ballerina.stdlib.http.compiler.Constants.CALLER_OBJ_NAME;
+import static io.ballerina.stdlib.http.compiler.Constants.COLON;
 import static io.ballerina.stdlib.http.compiler.Constants.DECIMAL;
 import static io.ballerina.stdlib.http.compiler.Constants.DECIMAL_ARRAY;
+import static io.ballerina.stdlib.http.compiler.Constants.DEFAULT;
 import static io.ballerina.stdlib.http.compiler.Constants.EMPTY;
 import static io.ballerina.stdlib.http.compiler.Constants.ERROR;
 import static io.ballerina.stdlib.http.compiler.Constants.FLOAT;
@@ -92,6 +105,7 @@ import static io.ballerina.stdlib.http.compiler.Constants.REQUEST_CONTEXT_OBJ_NA
 import static io.ballerina.stdlib.http.compiler.Constants.REQUEST_OBJ_NAME;
 import static io.ballerina.stdlib.http.compiler.Constants.RESOURCE_RETURN_TYPE;
 import static io.ballerina.stdlib.http.compiler.Constants.RESPONSE_OBJ_NAME;
+import static io.ballerina.stdlib.http.compiler.Constants.SERVICE_KEYWORD;
 import static io.ballerina.stdlib.http.compiler.Constants.STRING;
 import static io.ballerina.stdlib.http.compiler.Constants.STRING_ARRAY;
 import static io.ballerina.stdlib.http.compiler.Constants.STRUCTURED_ARRAY;
@@ -348,6 +362,47 @@ public final class HttpCompilerPluginUtil {
         return serviceDeclarationNode;
     }
 
+    public static ClassDefinitionNode getServiceClassDefinitionNode(SyntaxNodeAnalysisContext ctx) {
+        if (ctx.node().kind() != CLASS_DEFINITION) {
+            return null;
+        }
+        ClassDefinitionNode classDefinitionNode = (ClassDefinitionNode) ctx.node();
+        Optional<Symbol> serviceType = ctx.semanticModel().types()
+                .getTypeByName(BALLERINA, HTTP, EMPTY, HTTP_SERVICE_TYPE);
+        if (!hasServiceKeyWord(classDefinitionNode) || serviceType.isEmpty()) {
+            return null;
+        }
+        Optional<Symbol> symbol = ctx.semanticModel().symbol(classDefinitionNode);
+        if (symbol.isEmpty() || serviceType.get().kind() != TYPE_DEFINITION) {
+            return null;
+        }
+        ClassSymbol classSymbol = (ClassSymbol) symbol.get();
+        TypeSymbol serviceTypeSymbol = ((TypeDefinitionSymbol) serviceType.get()).typeDescriptor();
+        return classSymbol.subtypeOf(serviceTypeSymbol) ? classDefinitionNode : null;
+    }
+
+    public static AnnotationNode getAnnotationNode(SyntaxNodeAnalysisContext context) {
+        if (context.node().kind() != ANNOTATION) {
+            return null;
+        }
+        Optional<Symbol> symbol = context.semanticModel().symbol(context.node());
+        if (symbol.isPresent()) {
+            Optional<ModuleSymbol> module = symbol.get().getModule();
+            if (module.isEmpty() || !isHttpModule(module.get())) {
+                return null;
+            }
+            return (AnnotationNode) context.node();
+        }
+        // Added as a workaround for: https://github.com/ballerina-platform/ballerina-lang/issues/43525
+        return context.node().toSourceCode().trim().startsWith(AT_TOKEN.stringValue() + HTTP + COLON) ?
+                (AnnotationNode) context.node() : null;
+    }
+
+    private static boolean hasServiceKeyWord(ClassDefinitionNode classDefinitionNode) {
+        return classDefinitionNode.classTypeQualifiers()
+                .stream().anyMatch(token -> SERVICE_KEYWORD.equals(token.text().trim()));
+    }
+
     private static boolean isListenerBelongsToHttpModule(TypeSymbol listenerType) {
         if (listenerType.typeKind() == TypeDescKind.UNION) {
             return ((UnionTypeSymbol) listenerType).memberTypeDescriptors().stream()
@@ -378,13 +433,32 @@ public final class HttpCompilerPluginUtil {
             return false;
         }
 
-        Optional<Symbol> serviceContractType = semanticModel.types().getTypeByName(BALLERINA, HTTP, EMPTY,
+        Optional<Symbol> serviceType = semanticModel.types().getTypeByName(BALLERINA, HTTP, EMPTY,
                 HTTP_SERVICE_TYPE);
-        if (serviceContractType.isEmpty() ||
-                !(serviceContractType.get() instanceof TypeDefinitionSymbol serviceContractTypeDef)) {
+        if (serviceType.isEmpty() ||
+                !(serviceType.get() instanceof TypeDefinitionSymbol serviceTypeDef)) {
             return false;
         }
+        return serviceObjTypeDef.typeDescriptor().subtypeOf(serviceTypeDef.typeDescriptor());
+    }
 
-        return serviceObjTypeDef.typeDescriptor().subtypeOf(serviceContractTypeDef.typeDescriptor());
+    public static List<ResourceFunction> getResourceMethodWithDefaultAccessor(NodeList<Node> members) {
+        List<ResourceFunction> resourceFunctions = new ArrayList<>();
+        for (Node member : members) {
+            if (member.kind() != RESOURCE_ACCESSOR_DEFINITION && member.kind() != RESOURCE_ACCESSOR_DECLARATION) {
+                continue;
+            }
+            ResourceFunction resourceFunction = member.kind() == RESOURCE_ACCESSOR_DEFINITION
+                    ? new ResourceFunctionDefinition((FunctionDefinitionNode) member)
+                    : new ResourceFunctionDeclaration((MethodDeclarationNode) member);
+            if (DEFAULT.equalsIgnoreCase(resourceFunction.functionName().text().trim())) {
+                resourceFunctions.add(resourceFunction);
+            }
+        }
+        return resourceFunctions;
+    }
+
+    public static Document getDocument(SyntaxNodeAnalysisContext context) {
+        return context.currentPackage().module(context.moduleId()).document(context.documentId());
     }
 }